requirements.txt
httpx
supabase
pydantic
pydantic-ai
logfire
python-dotenv
fastapi
uvicorn
beautifulsoup4 